+++ util/coding.h
@@ -16,9 +16,9 @@
 namespace leveldb {
 
     void PutFixed32(std::string *dst, uint32_t value);
-    void PutFixed64(std::string *dst, uint32_t value);
+    void PutFixed64(std::string *dst, uint64_t value);
     void PutVarint32(std::string *dst, uint32_t value);
-    void PutVarint64(std::string *dst, uint32_t value);
+    void PutVarint64(std::string *dst, uint64_t value);
     void PutLengthPrefixedSlice(std::string *dst, const Slice &value);
 
     bool GetVarint32(Slice *input, uint32_t *value);
